Move podman specific code into PodmanUtils
Issue-Id: #24
